requirements.txt
<<<<<<< HEAD
crewai>=0.28.8
python-dotenv>=1.0.0
pydantic>=2.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.2.0
google-auth>=2.0.0
langchain-openai>=0.1.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0 
=======
crewai
langchain_google_genai
load_dotenv
crewai_tools
>>>>>>> 97a87092 (first commit)
